split utils and move errors
