+++ ansible_collections/arista/avd/plugins/filter/convert_schema.py
@@ -2,10 +2,10 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
+from ansible_collections.arista.avd.plugins.plugin_utils.errors import AristaAvdError
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdschema import AvdSchema
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdtodocumentationschemaconverter import AvdToDocumentationSchemaConverter
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdtojsonschemaconverter import AvdToJsonSchemaConverter
-from ansible_collections.arista.avd.plugins.plugin_utils.schema.errors import AristaAvdError
 
 
 def convert_schema(schema: dict, type: str):
