+++ ansible_collections/arista/avd/plugins/plugin_utils/errors/__init__.py
@@ -1,0 +1,8 @@
+from .errors import AristaAvdError, AristaAvdMissingVariableError, AvdSchemaError, AvdValidationError
+
+__all__ = [
+    "AristaAvdError",
+    "AristaAvdMissingVariableError",
+    "AvdSchemaError",
+    "AvdValidationError",
+]
