+++ ansible_collections/arista/avd/plugins/plugin_utils/schema/avdschema.py
@@ -2,10 +2,10 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
+from ansible_collections.arista.avd.plugins.plugin_utils.errors import AristaAvdError, AvdSchemaError, AvdValidationError
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avddataconverter import AvdDataConverter
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdschemaresolver import AvdSchemaResolver
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdvalidator import AVD_META_SCHEMA, AvdValidator
-from ansible_collections.arista.avd.plugins.plugin_utils.schema.errors import AristaAvdError, AvdSchemaError, AvdValidationError
 
 try:
     import jsonschema
