+++ ansible_collections/arista/avd/plugins/plugin_utils/utils/compile_searchpath.py
@@ -1,0 +1,31 @@
+from os.path import join as path_join
+
+
+def compile_searchpath(searchpath: list):
+    """
+    Create a new searchpath by inserting new items with <>/templates into the existing searchpath
+
+    This is copying the behavior of the "ansible.builtin.template" lookup module, and is necessary
+    to be able to load templates from all supported paths.
+
+    Example
+    -------
+    compile_searchpath(["patha", "pathb", "pathc"]) ->
+    ["patha", "patha/templates", "pathb", "pathb/templates", "pathc", "pathc/templates"]
+
+    Parameters
+    ----------
+    searchpath : list of str
+        List of Paths
+
+    Returns
+    -------
+    list of str
+        List of both original and extra paths with "/templates" added.
+    """
+
+    newsearchpath = []
+    for p in searchpath:
+        newsearchpath.append(path_join(p, "templates"))
+        newsearchpath.append(p)
+    return newsearchpath
