+++ ansible_collections/arista/avd/plugins/plugin_utils/utils/unique.py
@@ -1,0 +1,16 @@
+def unique(in_list):
+    """
+    Return list of unique items from the in_list
+
+    Parameters
+    ----------
+    in_list : list
+
+    Returns
+    -------
+    list
+        Unique list items
+    """
+
+    list_set = set(in_list)
+    return list(list_set)
