+++ ansible_collections/arista/avd/tests/unit/plugins/plugin_utils/schema/test_avdschema.py
@@ -8,8 +8,8 @@ import pytest
 import yaml
 from deepmerge import always_merger
 
+from ansible_collections.arista.avd.plugins.plugin_utils.errors import AvdValidationError
 from ansible_collections.arista.avd.plugins.plugin_utils.schema.avdschema import DEFAULT_SCHEMA, AvdSchema
-from ansible_collections.arista.avd.plugins.plugin_utils.schema.errors import AvdValidationError
 
 script_dir = os.path.dirname(__file__)
 with open(f"{script_dir}/access_lists.schema.yml", "r", encoding="utf-8") as schema_file:
